using long value for count in pager & easier access to utilty method

+++ cdmlib-services/src/main/java/eu/etaxonomy/cdm/api/service/pager/Pager.java
@@ -83,7 +83,7 @@ public interface Pager<T> {
      * If count > {@link #getPageSize()} then {@link #getPagesAvailable()} > 1
      * @return the total number of objects available.
      */
-    public Integer getCount();
+    public Long getCount();
 
     /**
      * Returns the index of the first record in this result set
